for subscription.pm and prorate.pm:
-modify the subscription and prorate price plans
(FS/FS/part_pkg/subscription.pm and prorate.pm) to have a
configurable (add a field to the %info hash) billing day instead of
"1st of the month" only.  subscription will be easy, prorate will be a
little trickier.

essentially, I replaced the '1' in the 'day' field of the timelocal that generates $$date with the value I added to the %info hash, 'cutoff_day'


-implement a price plan (new file in FS/FS/part_pkg/ - probably @ISA
FS::part_pkg::subscription) that charges the first full month if the
customer signs up between the 1st and the configurable billing day,
and gives them the remainder of the month free if they sign up
between the configurable billing day and the end of the month.

if this is the first time the customer is billed, and if the date is greater than the cutoff date, advance $ssdate to cutoff_day of next month, else $$date is cutoff_date of this month.  Either way, charge them for a month.

----------------------------------------------------------------------