ticket: 28287; add undiscounted monthly amount
I figured this need only parallel the discount report. That is, the discount
only is reported and applied if the period of recurrence is some whole number
of months. From my exploration of discounts it appeared that one could specify
discounts for other periods, but they did not appear on invoices.