Freeside sessionkey Should Not Be Stored in Plaintext
Add configurable option for storing the session key (in the
access_user_session table) as a salted SHA256 hash) based on setting
in the configuration.

If the 'hashsalt' configuration setting is a string of positive
length, use the salt string as input to the hash stored as the
sessionkey.

After setting the hashsalt, you will need to restart Apache,
freeside-queued, etc. You will also need to close and reopen your
browser and flush your cookies.